requirements.txt
pandas==2.2.2
pymongo==4.7.3
python-dotenv==1.0.1
statsmodels==0.14.4
tqdm==4.66.4
requests==2.32.3
xgboost==2.0.3
scikit-learn==1.6.1